Added launcher icon setup splash screen code.
The custom launcher icon was created for all resolutions and added to
the project. The code and layout for the splash screen is setup with it
being the first screen to be invoked and having the title bar disabled
and full screen. Need to still create and add splash screen image.